requirements.txt
python-telegram-bot>=21.0
yt-dlp>=2024.1.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
